Create v1.0 of GUI
This version is fully functional. It’s ugly, because matlab is a
hideous program. But it works well.

Current functionality:

Read in a certain csv file of collected data.
Based on user-selected values for fixed parameters, outputs multiple
spectra on a single plot for each non-fixed value.

For example, if we fix pot = 9 and brightness = 100, there will be 5
graphs corresponding to spectra for number = 0, 8, 16, 24, and 32, all
on the same plot.

Being able to use a slider to adjust values provides a great way to
visualize how changing a certain parameter affects the spectrum.

I’m so great at programming.